test: extract event cpi finding to a test crate
- fix clippy warnings
