+++ Cargo.toml
@@ -193,6 +193,7 @@ event-cpi = { path = "helpers/event-cpi" }
 event-cpi-macros = { path = "helpers/event-cpi-macros" }
 anchor-discriminators = { path = "helpers/anchor-discriminators" }
 anchor-discriminators-macros = { path = "helpers/anchor-discriminators-macros" }
+event-cpi-test-utils = { path = "helpers/event-cpi-test-utils" }
 evm-contracts-rs = { path = "helpers/evm-contracts-rs" }
 evm-contracts-test-suite = { path = "helpers/evm-contracts-test-suite" }
 gateway-event-stack = { path = "crates/gateway-event-stack", default-features = false }
