+++ programs/axelar-solana-gateway/tests/integration/approve_message.rs
@@ -14,9 +14,8 @@ use axelar_solana_gateway::{get_incoming_message_pda, get_validate_message_signi
 use axelar_solana_gateway_test_fixtures::gateway::{
     get_gateway_events, make_messages, make_verifier_set, random_message, GetGatewayError,
 };
-use axelar_solana_gateway_test_fixtures::{
-    assert_event_cpi, find_event_cpi, SolanaAxelarIntegration,
-};
+use axelar_solana_gateway_test_fixtures::SolanaAxelarIntegration;
+use event_cpi_test_utils::{assert_event_cpi, find_event_cpi};
 use itertools::Itertools;
 use pretty_assertions::assert_eq;
 use rand::Rng;
