+++ programs/dummy-axelar-solana-event-cpi/Cargo.toml
@@ -36,6 +36,7 @@ event-cpi.workspace = true
 event-cpi-macros.workspace = true
 
 [dev-dependencies]
+event-cpi-test-utils.workspace = true
 solana-program-test.workspace = true
 solana-sdk.workspace = true
 solana-client.workspace = true
